More chaos ensues. See message
Wondering about how to correctly handle events. There appear to be
multiple issues:
- Display: Traverse all widgets, draw them onto the canvas and show that
on screen. Drawing all widgets is only done once in the beginning, after
that we simply display the canvas because we update the canvas
immediately once an event occurs
- Handle Events: We cannot simply traverse the tree but must call the
relevant continuation. Therefore the new approach is directly incompatible
with the previous one. Instead of traversing, we have a datastructure in
which we find the relevant widget+continuation which we then update. In
theory this continuation doesn't even have to be associated with a
widget.
- Layouting: The layout has to be defined in code, ideally exactly where
we determine the functionality