Failed excursion, lessons learnt
effects have a troublesome property:
match effectful () with
| _ -> e1
| effect eff k -> ... continue k ... e2

All of: e1, continue k, e2
must have the same type as we don't know via which statement we will
return. F.e. continue might run effectful successfully to the end and
end up in e1, therefore returning the value of e1.
This also shows the second problem: Even when an effect occurs,
effectful () wil terminate at some point and return a value.

Therefore: effects are used !ONLY! for communication between the logic
and UI side. All event handling should be effectfree. Perhaps a monadic
point of view helps here: We call perform only after wrapping it in a UI
context which has already created the window. This perform is then free
to raise effects, end up in the UI, wait for events and then be resumed.
Then it is no longer problematic if the statement which is interrupted due to an effect
is completed but at a later point in time. This is exactly what we want!