Actual MVP using perform correctly
Some basic concepts used:
The UI is a tree of widgets.
An event is always handled by the root node which passes it to the
correct child etc. This can be seen as a distributed scheduler. Every
widget has a perform method which implements the logic. This is executed
once upon creation (i.e. the process of displaying the widget) and
terminates once the widget is no longer shown (either nominally by the
method terminating or by the parent throwing an exception if itself and
all children disappear (TODO))
While executing this perform the widget will normally perform an effect,
indicating that it is waiting for an event. To do so, it passes the
effect handler (normally the parent) the according trigger events and
its currently appearance which will remain unchanged until it handles an
event (Check if this actually works when implicitly handling resize
etc., perhaps it simply returns the list of triggers and provides a
display method)
This structure can be seen as a distributed scheduler. Upon receiving an
event the root widget is scheduled and distributes the event to the
correct child(ren). Then these children are scheduled, they perform
again and upon encountering another event requirement they return to the
parent via an effect.

TODO rename perform. An effect is performed, while self#perform does so
to, it also handles logic at the same time. Therefore a widget can
self#perform until it encounters a perform which is simply confusing.