Very large commit
Touches basically every part of the codebase.
The events have been abstracted from the SDL framework, we now use
polymorphic variants to represent events.
text_input has been reworked to use the new events and remove
unnecessary functions
Draw.ml has been looked at. Less complicated than I thought.
row.ml is still an issue. How do we handle global events like
keypresses? If two events receive it and both terminate we can't return
the values (as the type is 'a, not 'a list) This is a fundamental,
unsolvable problem
Solutions could be:
- Change type of row to 'a list
- Assume keypresses don't terminate widget (hardly possible)
- Allow only one widget to listen for keypresses. Works but is this
limited in some ways? It is natural for enter/text input to only work on
one widget but what about arrow keys etc? For *normal* UIs this should
probably not be a problem?